Dev branch merge (#24)
* 110 kV connection of substation

is now marked with connection_110kv = TRUE

* power lines deletion modification

power lines which are outside of Germany and are part of a relation
crossing the border are not deleted anymore.

* Bugfix

Conn_subgraphs was active but was not defined. Is now disabled.

* New result tables which indicate

 if a Substation is connected to the 110kV distribution grid. (In case
minimal voltage is above 110kV)

* Bugfix for deletion of open end lines

They haven’t been removed from power_circ_members even if the circuit
has been removed. Solved this issue now.

* Implemented connection of dead end relations with substations

also for Substations outside of Germany.